Refactor: Organize screens and navigation
