+++ app/src/main/java/com/example/caffeine/screen/coffeeSelectionScreen/CoffeeSelectionScreenUiState.kt
@@ -1,0 +1,8 @@
+package com.example.caffeine.screen.coffeeSelectionScreen
+
+import com.example.caffeine.R
+
+data class CoffeeSelectionScreenUiState(
+    val title: String = "Espresso",
+    val image: Int = R.drawable.espreesso,
+)
