+++ app/src/main/java/com/example/caffeine/screen/productScreen/ProductScreen.kt
@@ -1,32 +1,32 @@
-package com.example.caffeine.screen
+package com.example.caffeine.screen.productScreen
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Scaffold
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment.Companion.CenterHorizontally
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.unit.dp
 import androidx.navigation.NavController
-import com.example.caffeine.AppDestination
 import com.example.caffeine.R
-import com.example.caffeine.component.CoffeeSlider
-import com.example.caffeine.component.Header
+import com.example.caffeine.component.BackButtonHeader
+import com.example.caffeine.component.CoffeeSelector
+import com.example.caffeine.component.CupSize
 import com.example.caffeine.component.IconTextButton
-import com.example.caffeine.component.WelcomeMessage
-import com.example.caffeine.model.coffeeList
+import com.example.caffeine.component.SizeSelector
+import com.example.caffeine.navigation.AppDestination
 
 @Composable
-fun CoffeeSelectionScreen(
-    navController: NavController
+fun ProductScreen(
+    navController: NavController,
 ) {
-    val currentCoffeeTitle = remember { mutableStateOf(coffeeList.first().title) }
+    val args: String? = navController.currentBackStackEntry?.arguments?.getString("title")
+
     Scaffold(
         containerColor = Color.White
     ) { contentPadding ->
@@ -39,17 +39,22 @@ fun CoffeeSelectionScreen(
             horizontalAlignment = CenterHorizontally,
 
             ) {
-            Header(bottomSpace = 16.dp, modifier = Modifier.padding(horizontal = 16.dp))
-            WelcomeMessage()
-            CoffeeSlider(onSelectionChanged = {
-                currentCoffeeTitle.value = it
-            })
+            BackButtonHeader(
+                bottomSpace = 16.dp, modifier = Modifier.padding(horizontal = 16.dp),
+                title = checkNotNull(args)
+            ) {
+                navController.popBackStack()
+            }
+            CupSize()
+            SizeSelector()
+            Spacer(modifier = Modifier.height(16.dp))
+            CoffeeSelector()
             Spacer(modifier = Modifier.weight(1f))
             IconTextButton(
                 text = "Continue",
                 icon = painterResource(R.drawable.arrow_right),
                 onClick = {
-                    navController.navigate("${AppDestination.ProductScreen.route}/${currentCoffeeTitle.value}")
+                    navController.navigate(AppDestination.LoadingScreen.route)
                 }
             )
 
@@ -57,5 +62,7 @@ fun CoffeeSelectionScreen(
 
     }
 }
+
+
