+++ app/src/main/java/com/example/caffeine/screen/startScreen/StartScreen.kt
@@ -1,4 +1,4 @@
-package com.example.caffeine.screen
+package com.example.caffeine.screen.startScreen
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Column
@@ -12,11 +12,11 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.unit.dp
 import androidx.navigation.NavController
-import com.example.caffeine.AppDestination
 import com.example.caffeine.R
 import com.example.caffeine.component.Header
 import com.example.caffeine.component.IconTextButton
 import com.example.caffeine.component.MainText
+import com.example.caffeine.navigation.AppDestination
 
 @Composable
 fun StartScreen(
